% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockConditions.R
\name{add_condition,BlockConditions-method}
\alias{add_condition,BlockConditions-method}
\title{Add conditions}
\usage{
\S4method{add_condition}{BlockConditions}(object, variable, operator, value)
}
\description{
Add conditions
}
\keyword{internal}
