% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_filtering_transformator.R
\name{addCondition}
\alias{addCondition}
\title{Add a condition to a \code{BlockConditions} object}
\usage{
addCondition(object, variable, operator, value)
}
\arguments{
\item{object}{A \code{BlockConditions} object.}

\item{variable}{A character string specifying the variable/column name.}

\item{operator}{A character string specifying the operator (e.g., "==", "!=", "<", ">", "<=", ">=").}

\item{value}{The value to compare against.}
}
\value{
An updated \code{BlockConditions} object with the new condition added.
}
\description{
Add a condition to a \code{BlockConditions} object
}
\keyword{internal}
