% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_filtering_transformator_view_model.R
\name{filtering_transformator_model}
\alias{filtering_transformator_model}
\title{View model for \code{\link[=or_filtering_transformator]{or_filtering_transformator()}}.}
\description{
View model for \code{\link[=or_filtering_transformator]{or_filtering_transformator()}}.

View model for \code{\link[=or_filtering_transformator]{or_filtering_transformator()}}.
}
\keyword{Adds}
\keyword{Checks}
\keyword{alternative}
\keyword{an}
\keyword{conditions}
\keyword{duplicate}
\keyword{for}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{block_objects}}{Block objects.}

\item{\code{r_vals}}{Reactive values.}

\item{\code{alt_id}}{The id of the last alternative.}

\item{\code{block_conditions}}{Conditions as strings.}

\item{\code{choices_for_columns}}{Choices for the columns.}

\item{\code{final_filter_expr}}{Final filtering expression.}

\item{\code{final_exp}}{Final expression for \code{teal.data}.
Initializes the object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-filtering_transformator_model-new}{\code{filtering_transformator_model$new()}}
\item \href{#method-filtering_transformator_model-add_alternative}{\code{filtering_transformator_model$add_alternative()}}
\item \href{#method-filtering_transformator_model-is_duplicate_condition}{\code{filtering_transformator_model$is_duplicate_condition()}}
\item \href{#method-filtering_transformator_model-clone}{\code{filtering_transformator_model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filtering_transformator_model-new"></a>}}
\if{latex}{\out{\hypertarget{method-filtering_transformator_model-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{filtering_transformator_model$new(data, dataname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{the reactive data object from \code{teal}.}

\item{\code{dataname}}{\code{character(1)} the name of the dataset.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filtering_transformator_model-add_alternative"></a>}}
\if{latex}{\out{\hypertarget{method-filtering_transformator_model-add_alternative}{}}}
\subsection{Method \code{add_alternative()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{filtering_transformator_model$add_alternative()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisibly self.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filtering_transformator_model-is_duplicate_condition"></a>}}
\if{latex}{\out{\hypertarget{method-filtering_transformator_model-is_duplicate_condition}{}}}
\subsection{Method \code{is_duplicate_condition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{filtering_transformator_model$is_duplicate_condition(cond_str, block_cond_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cond_str}}{\code{character(1)} added condition.}

\item{\code{block_cond_list}}{\code{character(1)} list of all conditions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{logical(1)} whether the added condition is a duplicate.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filtering_transformator_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-filtering_transformator_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{filtering_transformator_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
