% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockConditions.R
\name{get_str_expression,BlockConditions-method}
\alias{get_str_expression,BlockConditions-method}
\title{Get condition expression}
\usage{
\S4method{get_str_expression}{BlockConditions}(object, dataname, data)
}
\description{
Get condition expression
}
\keyword{internal}
