% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockConditions.R
\name{get_str_expression}
\alias{get_str_expression}
\title{Get condition expression}
\usage{
get_str_expression(object, dataname, data)
}
\arguments{
\item{object}{A \code{BlockConditions} object.}

\item{dataname}{\code{character(1)} The name of the dataset to filter.}

\item{data}{The reactive \code{data} object from \code{teal}.}
}
\value{
\code{character(1)} The condition expression.
}
\description{
Get condition expression
}
\keyword{internal}
