% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_levels_transformator.R
\name{merge_level_transformer_ui}
\alias{merge_level_transformer_ui}
\title{UI design of the transformator}
\usage{
merge_level_transformer_ui(id)
}
\description{
UI design of the transformator
}
