% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_levels_transformator.R
\name{merge_levels_transformator}
\alias{merge_levels_transformator}
\title{Combine levels of a variable into one level}
\usage{
merge_levels_transformator(dataname, predefined = list())
}
\arguments{
\item{dataname}{(\code{character(1)}) the name of the dataset
which columns will be used for possible transformation.}

\item{predefined}{(\code{list}) the list which has variable name, levels and new label}
}
\value{
\code{teal::teal_transform_module}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This transformator allows the user to select a column from the dataset
and merge levels of this column into a new level. Only selected
levels are affected.
}
\examples{

app <- teal::init(
  data = teal.data::teal_data(IRIS = iris, code = "IRIS <- iris"),
  modules = teal::modules(
    teal::example_module(
      transformators = list(merge_levels_transformator(
        dataname = "IRIS",
        predefined = list(
          list("Species", "setosa", "SETOSA_WITHIN_FIX"),
          list("Petal.Width", c(0.2, 0.3, 0.5), 12)
        )
      ))
    )
  )
)
if (interactive()) {
  shiny::shinyApp(app$ui, app$server)
}

}
