% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchwork_plot_decorator.R
\name{patchwork_plot_decorator}
\alias{patchwork_plot_decorator}
\title{Patchwork Decorator}
\usage{
patchwork_plot_decorator(output_name, label_text = "decorator")
}
\arguments{
\item{output_name}{(\code{character(1)}) A name for the output plot object.}

\item{label_text}{(\code{character(1)}) A customised label text for the decorator.}
}
\value{
(\code{teal.data::qenv}) Returns a modified plot object with the transformation applied.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Decorator function to add plot title and footnote to patchwork plots
}
\details{
The module creates a UI with text controls for plot title and footnote.
The entered title and footnote text are applied to the patchwork plots.
}
