% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_report_manager.R
\name{report_manager_server}
\alias{report_manager_server}
\title{The server part of \code{\link[=tm_report_manager]{tm_report_manager()}}.}
\usage{
report_manager_server(id, reports_path = "reports", auto_save = TRUE, reporter)
}
\arguments{
\item{id}{(\code{character}) the id of the module.}

\item{reports_path}{character. Absolute path where reports should be stored.}

\item{auto_save}{logical. Whether to save active report whenever there are any changes made.}

\item{reporter}{the object that holds the report. Provided by \code{teal}.}
}
\description{
The server part of \code{\link[=tm_report_manager]{tm_report_manager()}}.
}
\keyword{internal}
