% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_report_manager.R
\name{report_manager_ui}
\alias{report_manager_ui}
\title{The \code{UI} of \code{\link[=tm_report_manager]{tm_report_manager()}}}
\usage{
report_manager_ui(id)
}
\arguments{
\item{id}{(\code{character}) the id of the module.}
}
\description{
The \code{UI} of \code{\link[=tm_report_manager]{tm_report_manager()}}
}
\keyword{internal}
