% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_report_manager.R
\name{tm_report_manager}
\alias{tm_report_manager}
\title{Report Manager Module}
\usage{
tm_report_manager(reports_path = "reports", auto_save = TRUE)
}
\arguments{
\item{reports_path}{character. Absolute path where reports should be stored.}

\item{auto_save}{logical. Whether to save active report whenever there are any changes made.}
}
\description{
Report Manager Module
}
\details{
This module supports collaborative work on \code{teal} reports, generated by \code{teal.reporter}
package. This module extends functionalities of \code{teal.reporter} and allows to store
reports in chosen path. Reports will be stored as JSON files inside folders named after
reports.
}
\examples{

app <- teal::init(
  data = teal.data::teal_data(IRIS = iris),
  modules = teal::modules(
    teal::example_module(transformators = list(or_filtering_transformator("IRIS"))),
    tm_report_manager()
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
