% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_levels_transformator.R
\name{TransformationManager}
\alias{TransformationManager}
\title{R6 Class for managing the transformation objects}
\description{
R6 Class for managing the transformation objects

R6 Class for managing the transformation objects
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TransformationManager-new}{\code{TransformationManager$new()}}
\item \href{#method-TransformationManager-add_id}{\code{TransformationManager$add_id()}}
\item \href{#method-TransformationManager-remove_id}{\code{TransformationManager$remove_id()}}
\item \href{#method-TransformationManager-reset}{\code{TransformationManager$reset()}}
\item \href{#method-TransformationManager-clone}{\code{TransformationManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransformationManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-TransformationManager-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransformationManager$new(predefined = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransformationManager-add_id"></a>}}
\if{latex}{\out{\hypertarget{method-TransformationManager-add_id}{}}}
\subsection{Method \code{add_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransformationManager$add_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransformationManager-remove_id"></a>}}
\if{latex}{\out{\hypertarget{method-TransformationManager-remove_id}{}}}
\subsection{Method \code{remove_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransformationManager$remove_id(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransformationManager-reset"></a>}}
\if{latex}{\out{\hypertarget{method-TransformationManager-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransformationManager$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransformationManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TransformationManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransformationManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
