% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uteals.R
\docType{package}
\name{uteals-package}
\alias{uteals}
\alias{uteals-package}
\title{Decorators and transformators for \code{teal} modules.}
\description{
Provides decorator and transformator modules
for \code{teal} modules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/phuse-org/uteals}
}

}
\author{
\strong{Maintainer}: Konrad Pagacz \email{konrad.pagacz@gmail.com}

Authors:
\itemize{
  \item  Johnson and Johnson \email{jnj@its.jnj.com}
}

}
