% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark_decorator.R
\name{watermark_decorator}
\alias{watermark_decorator}
\title{Watermark Decorator}
\usage{
watermark_decorator(output_name, watermark_text = "", font_size = 90)
}
\arguments{
\item{output_name}{(\code{character(1)}) a name for the output object (e.g., a plot or table).}

\item{watermark_text}{(\code{character(1)}) text to display for the watermark}

\item{font_size}{(\code{character(1)}) font size for the watermark text}
}
\value{
\code{\link[teal:teal_transform_module]{teal::teal_transform_module()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
A function to create a UI component for selecting watermark text
for tables or plots.
}
\details{
The module creates a UI with textInput for specifying watermark text and
font size.
the entered watermark text is displayed with a default gridify layout.
}
