% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_filtering_transformator.R
\name{addCondition,BlockConditions-method}
\alias{addCondition,BlockConditions-method}
\title{Add conditions}
\usage{
\S4method{addCondition}{BlockConditions}(object, variable, operator, value)
}
\description{
Add conditions
}
\keyword{internal}
