% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockConditions.R
\docType{class}
\name{BlockConditions-class}
\alias{BlockConditions-class}
\title{Class \code{BlockConditions}}
\description{
This class represents a collection of conditions used for filtering datasets.
}
\section{Slots}{

\describe{
\item{\code{conditions}}{A list of conditions,
where each condition is a list containing variable, operator, and value.}
}
}

\keyword{internal}
