% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportManager.R
\name{ReportManager}
\alias{ReportManager}
\title{\code{ReportManager} Object}
\description{
This can be used to manage state of report, load and save them into file
It will create \code{shiny::reactiveVal} variables and load report list from \code{report_path}
}
\keyword{internal}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{reports_path}}{(\code{character(1)})\cr
An absolute path to a folder where reports are stored}

    \item{\code{current_report_title}}{(\code{character(1)})\cr
\code{reactiveVal} field that stores current report title}

    \item{\code{available_reports}}{(\code{list()})\cr
\code{reactiveVal} field that stores a list of available reports}

    \item{\code{my_locked_report}}{(\code{character(1)})\cr
\code{reactiveVal} field that stores currently locked report title}

    \item{\code{read_only_mode}}{(\code{logical(1)})\cr
\code{reactiveVal} field that indicates if current report is in read-only mode}

    \item{\code{session}}{Placeholder for Shiny session object
Initialize \code{ReportManager}}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ReportManager-initialize}{\code{ReportManager$new()}}
    \item \href{#method-ReportManager-list_reports}{\code{ReportManager$list_reports()}}
    \item \href{#method-ReportManager-set_current_report_title}{\code{ReportManager$set_current_report_title()}}
    \item \href{#method-ReportManager-get_abs_report_path}{\code{ReportManager$get_abs_report_path()}}
    \item \href{#method-ReportManager-save_report_files}{\code{ReportManager$save_report_files()}}
    \item \href{#method-ReportManager-load_report}{\code{ReportManager$load_report()}}
    \item \href{#method-ReportManager-merge_reports}{\code{ReportManager$merge_reports()}}
    \item \href{#method-ReportManager-reset}{\code{ReportManager$reset()}}
    \item \href{#method-ReportManager-delete_report}{\code{ReportManager$delete_report()}}
    \item \href{#method-ReportManager-is_locked_by_other}{\code{ReportManager$is_locked_by_other()}}
    \item \href{#method-ReportManager-rename_report}{\code{ReportManager$rename_report()}}
    \item \href{#method-ReportManager-is_new_report}{\code{ReportManager$is_new_report()}}
    \item \href{#method-ReportManager-create_new_report}{\code{ReportManager$create_new_report()}}
    \item \href{#method-ReportManager-auto_save_observer}{\code{ReportManager$auto_save_observer()}}
    \item \href{#method-ReportManager-unlock_report_public}{\code{ReportManager$unlock_report_public()}}
    \item \href{#method-ReportManager-release_lock}{\code{ReportManager$release_lock()}}
    \item \href{#method-ReportManager-setup_csv_download}{\code{ReportManager$setup_csv_download()}}
    \item \href{#method-ReportManager-export_tables_to_csv}{\code{ReportManager$export_tables_to_csv()}}
    \item \href{#method-ReportManager-rebuild_report}{\code{ReportManager$rebuild_report()}}
    \item \href{#method-ReportManager-clone}{\code{ReportManager$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-initialize}{}}}
\subsection{\code{ReportManager$new()}}{
  This can be used to manage state of report, load and save them into file
It will create \code{shiny::reactiveVal} variables and load report list from \code{reports_path}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$new(reports_path, session)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{reports_path}}{(character) An absolute path to where reports are stored}
      \item{\code{session}}{Session object passed from \code{moduleServer}}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{ReportManager} object
List available reports from objects' \code{reports_path}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-list_reports"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-list_reports}{}}}
\subsection{\code{ReportManager$list_reports()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$list_reports()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{data.frame} with fields \code{reports}, \code{created_by}, \code{locked_by}, and \code{last_rebuild}
Method for storing loaded report title inside object.
This assigns a value in \code{current_report_title} field
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-set_current_report_title"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-set_current_report_title}{}}}
\subsection{\code{ReportManager$set_current_report_title()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$set_current_report_title(title_or_index)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{title_or_index}}{(character/numeric) Title of report or an index from objects' \code{available_reports}
Get absolute report path}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-get_abs_report_path"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-get_abs_report_path}{}}}
\subsection{\code{ReportManager$get_abs_report_path()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$get_abs_report_path(title)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{title}}{(character) A title of report to get path of}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    (character) A path of desired report
Save report with provided title
This utilizes methods from \code{teal.reporter::Reporter} to store files required to load and save reports
If a report with specific title exists, it will overwrite it.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-save_report_files"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-save_report_files}{}}}
\subsection{\code{ReportManager$save_report_files()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$save_report_files(report_title, reporter, first = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{report_title}}{(\code{character}) A title for the saved report.}
      \item{\code{reporter}}{(\code{Reporter}) A reporter object passed from \code{teal::init}}
      \item{\code{first}}{(\code{logical}) Whether this is the first time saving the report.

Load report to report previewer
This utilizes methods from \code{teal.reporter::Reporter} to load reports}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-load_report"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-load_report}{}}}
\subsection{\code{ReportManager$load_report()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$load_report(selected_title, reporter, skip_lock = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{selected_title}}{(\code{integer} or \code{character}) An index or a title of a report to load}
      \item{\code{reporter}}{(\code{Reporter}) A reporter object passed from \code{teal::init}}
      \item{\code{skip_lock}}{(\code{boolean}) Whether to skip locking a report or not
Merge reports together}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-merge_reports"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-merge_reports}{}}}
\subsection{\code{ReportManager$merge_reports()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$merge_reports(x_title, y_title, reporter)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{x_title}}{Base report title}
      \item{\code{y_title}}{Title of report to merge into x}
      \item{\code{reporter}}{A \code{TealReporter} object
Reset report manager}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-reset"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-reset}{}}}
\subsection{\code{ReportManager$reset()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$reset()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    the return of \code{self$list_reports()}
Delete report
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-delete_report"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-delete_report}{}}}
\subsection{\code{ReportManager$delete_report()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$delete_report(selected_title)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{selected_title}}{(integer or character) An index or a title of a report to load
Checks if report is locked by another user}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-is_locked_by_other"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-is_locked_by_other}{}}}
\subsection{\code{ReportManager$is_locked_by_other()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$is_locked_by_other(
  report_title,
  verbose = TRUE,
  message_type = "error"
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{report_title}}{(character) A title for the saved report.}
      \item{\code{verbose}}{(boolean) Whether to show notification on locked report or not}
      \item{\code{message_type}}{(character) \verb{shiny::showNotification's} type argument}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    logical(1) TRUE if locked by another user, FALSE if unlocked or locked by me
Rename report
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-rename_report"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-rename_report}{}}}
\subsection{\code{ReportManager$rename_report()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$rename_report(old_title, new_title, reporter)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{old_title}}{(character) Current report title}
      \item{\code{new_title}}{(character) New report title}
      \item{\code{reporter}}{(Reporter) A reporter object
Check if this would be a new report}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-is_new_report"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-is_new_report}{}}}
\subsection{\code{ReportManager$is_new_report()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$is_new_report(reporter)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{reporter}}{(Reporter) A reporter object}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    logical indicating if this is a new report
Create new report with title
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-create_new_report"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-create_new_report}{}}}
\subsection{\code{ReportManager$create_new_report()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$create_new_report(report_title, reporter)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{report_title}}{(character) Title for the new report}
      \item{\code{reporter}}{(Reporter) A reporter object
Add \code{observeEvent} to invoke auto save on reporter cards change.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-auto_save_observer"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-auto_save_observer}{}}}
\subsection{\code{ReportManager$auto_save_observer()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$auto_save_observer(reporter)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{reporter}}{(Reporter) A reporter object passed from \code{teal::init}
Unlock a locked report}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-unlock_report_public"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-unlock_report_public}{}}}
\subsection{\code{ReportManager$unlock_report_public()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$unlock_report_public(report_title)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{report_title}}{(character) Title of the report to unlock
Release lock on current report (keep report loaded but make it read-only)}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-release_lock"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-release_lock}{}}}
\subsection{\code{ReportManager$release_lock()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$release_lock(report_title = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{report_title}}{(character) Title of the report to release lock from
Set up CSV download observer}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-setup_csv_download"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-setup_csv_download}{}}}
\subsection{\code{ReportManager$setup_csv_download()}}{
  Registers the \code{observeEvent} that handles CSV download clicks.
Call once from \code{moduleServer} after initialisation.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$setup_csv_download(input)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{input}}{Shiny input object from \code{moduleServer}
Export all tables from a saved report as CSV files}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-export_tables_to_csv"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-export_tables_to_csv}{}}}
\subsection{\code{ReportManager$export_tables_to_csv()}}{
  Loads the report from disk, extracts all table-type content from each card,
converts each to a \code{data.frame} and writes it as a CSV file in \code{output_dir}.
Returns the vector of written file paths, or an empty character vector when
no tables are found.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$export_tables_to_csv(report_title, output_dir)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{report_title}}{(character) Title of the report to export.}
      \item{\code{output_dir}}{(character) Directory where CSV files will be written.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    character vector of written file paths.
Re-build reports
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-rebuild_report"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-rebuild_report}{}}}
\subsection{\code{ReportManager$rebuild_report()}}{
  Rebuild reports to include data that has changed.
This will replace loaded report cards with new cards, that were rebuilt from
a code used to generate these cards.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$rebuild_report(report_title, rp)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{report_title}}{(character) A title for the saved report.}
      \item{\code{rp}}{(Reporter) A reporter object passed from \code{teal::init}
Get current user name
Get metadata file path
Lock report so that it can't be overwritten by another user
Unlock report
Save creator information
Save code from each card as \code{code.rds} in the report directory.
Extract all table-type content from a Reporter as a named list of data.frames
Register \code{onSessionEnded} to unlock report when session is closed}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReportManager-clone}{}}}
\subsection{\code{ReportManager$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ReportManager$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
