% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uteals.R
\docType{package}
\name{uteals-package}
\alias{uteals}
\alias{uteals-package}
\title{Decorators and transformators for \code{teal} modules.}
\description{
Provides decorator and transformator modules
for \code{teal} modules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/phuse-org/uteals}
}

}
\author{
\strong{Maintainer}: Peyman Eshghi \email{peymaan.es@gmail.com}

Authors:
\itemize{
  \item Nadia Abraham \email{NAbraha5@ITS.JNJ.com}
  \item Chanchal Bhalla \email{CBhalla@ITS.JNJ.com}
  \item Pranith Gourisetty \email{PGourise@ITS.JNJ.com}
  \item Sohan Lal \email{SLal8@ITS.JNJ.com}
  \item Przemyslaw Posiadala \email{PPosiada@ITS.JNJ.com}
  \item Alina Tselinina \email{ATselini@ITS.JNJ.com}
  \item Konrad Pagacz \email{konrad.pagacz@gmail.com}
}

Other contributors:
\itemize{
  \item PHUSE [copyright holder]
}

}
